refactor: 운세 모델 로딩 및 추론 로직 계층별 분리
